+++ just_audio/android/src/main/java/com/ryanheise/just_audio/AudioPlayer.java
@@ -96,6 +96,7 @@ public class AudioPlayer implements MethodCallHandler, Player.Listener, Metadata
     private AudioAttributes pendingAudioAttributes;
     private LoadControl loadControl;
     private boolean offloadSchedulingEnabled;
+    private AudioOffloadPreferences audioOffloadPreferences;
     private boolean useLazyPreparation;
     private LivePlaybackSpeedControl livePlaybackSpeedControl;
     private List<Object> rawAudioEffects;
@@ -145,6 +146,7 @@ public class AudioPlayer implements MethodCallHandler, Player.Listener, Metadata
         final String id,
         Map<?, ?> audioLoadConfiguration,
         List<Object> rawAudioEffects,
+        Map<?, ?> audioOffloadPreferences,
         Boolean offloadSchedulingEnabled,
         boolean useLazyPreparation
     ) {
@@ -152,6 +154,24 @@ public class AudioPlayer implements MethodCallHandler, Player.Listener, Metadata
         this.rawAudioEffects = rawAudioEffects;
         this.offloadSchedulingEnabled = offloadSchedulingEnabled != null ? offloadSchedulingEnabled : false;
         this.useLazyPreparation = useLazyPreparation;
+
+        if (audioOffloadPreferences != null) {
+            this.audioOffloadPreferences = new AudioOffloadPreferences.Builder()
+                .setIsGaplessSupportRequired((Boolean)audioOffloadPreferences.get("isGaplessSupportRequired"))
+                .setIsSpeedChangeSupportRequired((Boolean)audioOffloadPreferences.get("isSpeedChangeSupportRequired"))
+                .setAudioOffloadMode((Integer)audioOffloadPreferences.get("audioOffloadMode"))
+                .build();
+        } else {
+            final int offloadMode = offloadSchedulingEnabled
+                ? AudioOffloadPreferences.AUDIO_OFFLOAD_MODE_ENABLED
+                : AudioOffloadPreferences.AUDIO_OFFLOAD_MODE_DISABLED;
+            this.audioOffloadPreferences = new AudioOffloadPreferences.Builder()
+                .setIsGaplessSupportRequired(!offloadSchedulingEnabled)
+                .setIsSpeedChangeSupportRequired(!offloadSchedulingEnabled)
+                .setAudioOffloadMode(offloadMode)
+                .build();
+        }
+
         methodChannel = new MethodChannel(messenger, "com.ryanheise.just_audio.methods." + id);
         methodChannel.setMethodCallHandler(this);
         eventChannel = new BetterEventChannel(messenger, "com.ryanheise.just_audio.events." + id);
@@ -772,27 +792,10 @@ public class AudioPlayer implements MethodCallHandler, Player.Listener, Metadata
                 builder.setLivePlaybackSpeedControl(livePlaybackSpeedControl);
             }
             player = builder.build();
-            // The latest ExoPlayer enables offload scheduling by default but
-            // it doesn't support gapless playback below SDK level 33 or speec
-            // changing. To maintain backwards compatibility within just_audio,
-            // we set gapless playback and speed changing support as required,
-            // and let ExoPlayer choose whether it can enable offload
-            // scheduling depending on device support. If the app passes in
-            // androidOffloadSchedulingEnabled: true, we simply remove these
-            // requirements which may prevent gapless and speed changing from
-            // working, but will allow offload to work. A future release may
-            // expose more parameters to the app concerning offloading
-            // preferences.
             player.setTrackSelectionParameters(
                 player.getTrackSelectionParameters()
                     .buildUpon()
-                    .setAudioOffloadPreferences(
-                        new AudioOffloadPreferences.Builder()
-                            .setIsGaplessSupportRequired(!offloadSchedulingEnabled)
-                            .setIsSpeedChangeSupportRequired(!offloadSchedulingEnabled)
-                            .setAudioOffloadMode(AudioOffloadPreferences.AUDIO_OFFLOAD_MODE_ENABLED)
-                            .build()
-                    )
+                    .setAudioOffloadPreferences(audioOffloadPreferences)
                     .build()
             );
             setAudioSessionId(player.getAudioSessionId());