+++ just_audio/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.10.5
+
+* Disable Android audio offload by default to prevent playback issues.
+* Add androidAudioOffloadPreferences.
+* Deprecate androidOffloadSchedulingEnabled.
+
 ## 0.10.4
 
 * Fix bug on simultaneous loads.